#2691 Moved IUserPreferencesProvider to pojo-bl.
