+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/web/interfaces/IUserPreferencesProvider.java
@@ -1,0 +1,26 @@
+package ua.com.fielden.platform.web.interfaces;
+
+import com.google.inject.ImplementedBy;
+import ua.com.fielden.platform.security.user.User;
+import ua.com.fielden.platform.web.interfaces.impl.NoOptUserPreferencesProvider;
+
+import java.util.Map;
+
+/// A contract for providing user-specific application preferences delivered to a web client as part of the application configuration.
+///
+/// The returned key/value pairs are merged into the application configuration response sent to the frontend,
+/// on top of the standard platform-level configuration provided by `ApplicationConfigurationResource`.
+/// Entries with keys matching the standard configuration keys (e.g., `locale`, `dateFormat`, `timeZone`) will override the default values.
+///
+/// Implement and bind this interface in an application-specific IoC module to provide user-specific preferences.
+///
+@FunctionalInterface
+@ImplementedBy(NoOptUserPreferencesProvider.class)
+public interface IUserPreferencesProvider {
+
+    /// Returns a map of application settings (keys) and their values specific to `user`.
+    /// Entries with keys matching the standard platform-level application settings will override the defaults.
+    ///
+    Map<String, Object> getPreferencesFor(final User user);
+
+}
