+++ platform-web-resources/src/main/java/ua/com/fielden/platform/web/resources/webui/ApplicationConfigurationResource.java
@@ -12,7 +12,7 @@ import ua.com.fielden.platform.basic.config.IApplicationSettings;
 import ua.com.fielden.platform.security.user.IUserProvider;
 import ua.com.fielden.platform.utils.IDates;
 import ua.com.fielden.platform.web.app.IWebUiConfig;
-import ua.com.fielden.platform.web.application.IUserPreferencesProvider;
+import ua.com.fielden.platform.web.interfaces.IUserPreferencesProvider;
 import ua.com.fielden.platform.web.interfaces.IDeviceProvider;
 import ua.com.fielden.platform.web.resources.RestServerUtil;
